algorithm-leetcode/src/test/com/hanxian/algorithm/leetcode/daily/questions/Solution1457TwiceTest.java


package com.hanxian.algorithm.leetcode.daily.questions;

import com.hanxian.algorithm.leetcode.common.TreeNode;import org.mockito.junit.jupiter.MockitoExtension;
import static org.mockito.MockitoAnnotations.openMocks;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.spy;
import java.util.concurrent.Callable;
import java.util.concurrent.CompletableFuture;
import java.util.concurrent.Future;
import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.assertj.core.api.Assertions.within;
    import static org.mockito.ArgumentMatchers.any;
    import static org.mockito.ArgumentMatchers.anyInt;
    import static org.mockito.ArgumentMatchers.anyString;
    import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.doAnswer;
import static org.mockito.Mockito.doReturn;
import static org.mockito.Mockito.doThrow;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.mockito.InjectMocks;
import org.mockito.stubbing.Answer;

class Solution1457TwiceTest {

    private Solution1457Twice solution1457TwiceUnderTest;

@BeforeEach
void setUp() throws Exception {
            solution1457TwiceUnderTest = new Solution1457Twice() ;
}
                
    @Test
    void testPseudoPalindromicPaths() throws Exception {
    // Setup
                        final TreeNode root = new TreeNode(0, null, null);

    // Run the test
 final int result =  solution1457TwiceUnderTest.pseudoPalindromicPaths(root);

        // Verify the results
 assertThat(result).isEqualTo( 0 ) ;
    }
                        }

